compile now also exports line metadata
This is currently exporting as a standalone CSV. It could be merged into the line one later if we want.
I think it is fine being standalone so that games that don't need it don't have to care about it.
The metadata does technically duplicate the lineId in both the data and as the row key but that's also fine I reckon, and easy enough to fix later on if we want.